Enhance CampaignAttachment with status badge and edit functionality
- Integrated CampaignDrawer for editing campaign details, improving user interaction.
- Added a status badge to visually represent campaign schedule status, enhancing clarity.
- Refactored layout to maintain consistency and improve user experience with content uploads and campaign details.
- Updated content display logic to accommodate new features and ensure better feedback for users.

+++ components/content/content-result-chart-config.ts
@@ -56,5 +56,5 @@ export function formatControlMonth(month: string) {
   if (!year || !monthPart) return month
   const date = new Date(Number(year), Number(monthPart) - 1, 1)
   if (Number.isNaN(date.getTime())) return month
-  return date.toLocaleDateString("en-US", { month: "short" })
+  return date.toLocaleDateString("en-US", { month: "long" })
 }
